// src/app/core/interfaces/patient-register-interface.ts
export interface PatientRegisterInterface {
  nationalId: string;
  fName: string;
  lName: string;
  userName: string;
  email: string;
  password: string;
  gender: number;
  country: string;
  city: string;
  age: number;
  currentStage: number;
  height: number;
  weight: number;
  bloodType: number;
  chronicDisease: string;
}